evolution chain script

+++ database/enhancer.js
@@ -1,62 +1,38 @@
 // for each pokemon from id 1 to 151 add a new json related to enhanced data (color, evolution stage, habitat)
 
 import axios from "axios";
-import { writeFileSync } from "fs";
+import { writeFileSync, readFileSync, existsSync } from "fs";
 import { join } from "path";
 
-const getPokemonFromEvolutionChain = async (evolutionChainId) => {
-  const evolutionChain = await axios.get(
-    `https://pokeapi.co/api/v2/evolution-chain/${evolutionChainId}`
-  );
-  const { data } = evolutionChain;
-  const pokemonEvolutions = [];
-  const getPokemon = (chain, stage) => {
-    const id = chain.species.url.split("/")[6];
-    if (id <= 151 && chain.species)
-      pokemonEvolutions.push({
-        id: Number(id),
-        name: chain.species.name,
-        stage,
-      });
-
-    if (chain.evolves_to) {
-      chain.evolves_to.forEach((evolution) => {
-        getPokemon(evolution, id <= 151 ? stage + 1 : stage);
-      });
-    }
-  };
-  getPokemon(data.chain, 1);
-  return pokemonEvolutions;
-};
-
-const gatherEvolutionChainInfo = async () => {
-  const promises = [];
-  for (let i = 1; i <= 78; i++) promises.push(getPokemonFromEvolutionChain(i));
-
-  const pokemonEvolutions = await Promise.all(promises);
-  return pokemonEvolutions.flat();
-};
+const NUMBER_OF_POKEMON = 1025;
 
 const enhancer = async () => {
   const pokeData = [];
-  const pokemonEvolutionData = await gatherEvolutionChainInfo();
-  for (let i = 1; i <= 151; i++) {
+
+  const evolutionData = JSON.parse(
+    readFileSync(join("./data/", `evolution-data.json`), "utf8")
+  );
+
+  const pokemonEvolutionData = evolutionData.filter(Boolean);
+
+  for (let i = 1; i <= NUMBER_OF_POKEMON; i++) {
+    // if the file already exists, skip it
+    const checkFilePath = join("./data/enhanced", `pokemon-enhanced-${i}.json`);
+    if (existsSync(checkFilePath, "utf8")) continue;
+
     const response = await axios.get(`https://pokeapi.co/api/v2/pokemon/${i}`);
     const { data } = response;
     const speciesResponse = await axios.get(data.species.url);
     const { data: speciesData } = speciesResponse;
     const color = speciesData.color.name;
-    const evolutionStage = pokemonEvolutionData.find(
-      (poke) => poke.id === i
-    ).stage;
+    const evolutionStage =
+      pokemonEvolutionData.find((poke) => poke.id === i)?.stage ?? 1;
     const habitat = speciesData.habitat ? speciesData.habitat.name : "unknown";
     const enhancedData = { id: data.id, color, evolutionStage, habitat };
     pokeData.push(enhancedData);
-    const filePath = join("./data/", `pokemon-enhanced-${i}.json`);
+    const filePath = join("./data/enhanced/", `pokemon-enhanced-${i}.json`);
     writeFileSync(filePath, JSON.stringify(enhancedData));
   }
-  const filePath = join("./data/", "pokemon-enhanced.json");
-  writeFileSync(filePath, JSON.stringify(pokeData));
 };
 
 enhancer();
